@@ -1,19 +1,21 @@
 #pragma once
 
-#include "fmt/core.h"
 #include "sme/image_stack.hpp"
-#include "sme_exception.hpp"
 #include <QImage>
 #include <algorithm>
-#include <pybind11/numpy.h>
-#include <pybind11/stl.h>
+#include <fmt/core.h>
+#include <nanobind/make_iterator.h>
+#include <nanobind/nanobind.h>
+#include <nanobind/ndarray.h>
 #include <stdexcept>
 #include <vector>
 
-namespace sme {
+namespace pysme {
 
-pybind11::array toPyImageRgb(const sme::common::ImageStack &imgs);
-pybind11::array toPyImageMask(const sme::common::ImageStack &imgs);
+nanobind::ndarray<nanobind::numpy, std::uint8_t>
+toPyImageRgb(const ::sme::common::ImageStack &imgs);
+nanobind::ndarray<nanobind::numpy, bool>
+toPyImageMask(const ::sme::common::ImageStack &imgs);
 
 template <typename T> std::string vecToNames(const std::vector<T> &vec) {
   std::string str;
@@ -30,7 +32,7 @@ T &findElem(std::vector<T> &v, typename std::vector<T>::difference_type i) {
   }
   auto idx{static_cast<std::size_t>(i)};
   if (i < 0 || idx >= v.size()) {
-    throw sme::SmeInvalidArgument(fmt::format("index {} out of bounds", i));
+    throw std::invalid_argument(fmt::format("index {} out of bounds", i));
   }
   return v[idx];
 }
@@ -41,56 +43,64 @@ template <typename T> T &findElem(std::vector<T> &v, const std::string &name) {
       iter != v.end()) {
     return *iter;
   }
-  throw SmeInvalidArgument(fmt::format("name '{}' not found", name));
+  throw std::invalid_argument(fmt::format("name '{}' not found", name));
 }
 
 // a read-only iterable list of T with index or name-based lookup
-template <typename T> void bindList(pybind11::module &m, const char *typeName) {
+template <typename T>
+void bindList(nanobind::module_ &m, const char *typeName) {
   std::string listName{fmt::format("{}List", typeName)};
   std::string docString{
       fmt::format("a list of :class:`{1}` objects\n\n"
                   "the list can be iterated over, or an "
                   "element can be looked up by its index or name",
                   listName, typeName)};
-  pybind11::class_<std::vector<T>>(m, listName.c_str(), docString.c_str())
-      .def(pybind11::init<>())
+  nanobind::class_<std::vector<T>>(m, listName.c_str(), docString.c_str())
+      .def(nanobind::init<>())
       .def("__len__", [](const std::vector<T> &v) { return v.size(); })
       .def(
           "__getitem__",
           [](std::vector<T> &v, typename std::vector<T>::difference_type i)
-              -> T & { return sme::findElem(v, i); },
-          pybind11::return_value_policy::reference_internal)
+              -> T & { return findElem(v, i); },
+          nanobind::rv_policy::reference_internal)
       .def(
           "__getitem__",
           [](std::vector<T> &v, const std::string &name) -> T & {
-            return sme::findElem(v, name);
+            return findElem(v, name);
           },
-          pybind11::return_value_policy::reference_internal)
+          nanobind::rv_policy::reference_internal)
       .def(
           "__iter__",
           [](std::vector<T> &v) {
-            return pybind11::make_iterator(v.begin(), v.end());
+            return nanobind::make_iterator(
+                nanobind::type<T>(), "sme_list_iterator", v.begin(), v.end());
           },
-          pybind11::keep_alive<0, 1>());
+          nanobind::keep_alive<0, 1>());
 }
 
-// return the contents of a std::vector as an ndarray, with optional shape,
+// return the contents of a std::vector as a numpy ndarray with a given shape,
 // without making a copy of the data, based on
 // https://github.com/pybind/pybind11/issues/1042#issuecomment-642215028
-template <typename T>
-static inline pybind11::array_t<T>
-as_ndarray(std::vector<T> &&v, const std::vector<ssize_t> &shape = {}) {
+template <typename UnderlyingType, typename Type = UnderlyingType>
+static inline nanobind::ndarray<nanobind::numpy, Type>
+as_ndarray(std::vector<UnderlyingType> &&v,
+           std::initializer_list<std::size_t> shape) {
   auto data{v.data()};
-  auto size{static_cast<ssize_t>(v.size())};
-  auto ptr{std::make_unique<std::vector<T>>(std::move(v))};
-  auto capsule{pybind11::capsule(ptr.get(), [](void *p) {
-    std::unique_ptr<std::vector<T>>(reinterpret_cast<std::vector<T> *>(p));
-  })};
+  auto ptr = std::make_unique<std::vector<UnderlyingType>>(std::move(v));
+  auto capsule = nanobind::capsule(ptr.get(), [](void *p) noexcept {
+    std::unique_ptr<std::vector<UnderlyingType>>(
+        static_cast<std::vector<UnderlyingType> *>(p));
+  });
   ptr.release();
-  if (shape.empty()) {
-    return pybind11::array(size, data, capsule);
-  }
-  return pybind11::array(shape, data, capsule);
+  return nanobind::ndarray<nanobind::numpy, Type>(data, shape, capsule);
 }
 
-} // namespace sme
+template <typename UnderlyingType, typename Type = UnderlyingType>
+static inline nanobind::ndarray<nanobind::numpy, Type>
+as_ndarray(std::vector<UnderlyingType> &&v,
+           const ::sme::common::Volume &shape) {
+  return as_ndarray<UnderlyingType, Type>(
+      std::move(v), {shape.depth(), static_cast<std::size_t>(shape.height()),
+                     static_cast<std::size_t>(shape.width())});
+}
+} // namespace pysme
