@@ -1,34 +1,36 @@
-// Python.h (included by pybind11.h) must come first
+// Python.h (#included by nanobind.h) must come first
 // https://docs.python.org/3.2/c-api/intro.html#include-files
-#include <pybind11/pybind11.h>
+#include <nanobind/nanobind.h>
 
 #include "sme/logger.hpp"
 #include "sme/model.hpp"
 #include "sme_common.hpp"
 #include "sme_compartment.hpp"
+#include <nanobind/stl/string.h>
 
-namespace sme {
+namespace pysme {
 
-void pybindCompartment(pybind11::module &m) {
-  sme::bindList<Compartment>(m, "Compartment");
-  pybind11::class_<Compartment>(m, "Compartment",
+void bindCompartment(nanobind::module_ &m) {
+  bindList<Compartment>(m, "Compartment");
+  nanobind::class_<Compartment>(m, "Compartment",
                                 R"(
                                 a compartment where species live
                                 )")
-      .def_property("name", &Compartment::getName, &Compartment::setName,
-                    R"(
+      .def_prop_rw("name", &Compartment::getName, &Compartment::setName,
+                   R"(
                     str: the name of this compartment
                     )")
-      .def_readonly("species", &Compartment::species,
-                    R"(
+      .def_ro("species", &Compartment::species,
+              R"(
                     SpeciesList: the species in this compartment
                     )")
-      .def_readonly("reactions", &Compartment::reactions,
-                    R"(
+      .def_ro("reactions", &Compartment::reactions,
+              R"(
                     ReactionList: the reactions in this compartment
                     )")
-      .def_readonly("geometry_mask", &Compartment::geometry_mask,
-                    R"(
+      .def_prop_ro("geometry_mask", &Compartment::geometry_mask,
+                   nanobind::rv_policy::take_ownership,
+                   R"(
                     numpy.ndarray: a voxel mask of the compartment geometry
 
                     An array of boolean values, where
@@ -60,7 +62,8 @@ void pybindCompartment(pybind11::module &m) {
       .def("__str__", &Compartment::getStr);
 }
 
-Compartment::Compartment(model::Model *sbmlDocWrapper, const std::string &sId)
+Compartment::Compartment(::sme::model::Model *sbmlDocWrapper,
+                         const std::string &sId)
     : s(sbmlDocWrapper), id(sId) {
   const auto &compSpecies = s->getSpecies().getIds(id.c_str());
   species.reserve(static_cast<std::size_t>(compSpecies.size()));
@@ -72,12 +75,6 @@ Compartment::Compartment(model::Model *sbmlDocWrapper, const std::string &sId)
       reactions.emplace_back(s, reac.toStdString());
     }
   }
-  updateMask();
-}
-
-void Compartment::updateMask() {
-  geometry_mask = toPyImageMask(
-      s->getCompartments().getCompartment(id.c_str())->getCompartmentImages());
 }
 
 std::string Compartment::getName() const {
@@ -88,6 +85,11 @@ void Compartment::setName(const std::string &name) {
   s->getCompartments().setName(id.c_str(), name.c_str());
 }
 
+nanobind::ndarray<nanobind::numpy, bool> Compartment::geometry_mask() const {
+  return toPyImageMask(
+      s->getCompartments().getCompartment(id.c_str())->getCompartmentImages());
+}
+
 std::string Compartment::getStr() const {
   std::string str("<sme.Compartment>\n");
   str.append(fmt::format("  - name: '{}'\n", getName()));
@@ -95,30 +97,4 @@ std::string Compartment::getStr() const {
   return str;
 }
 
-} // namespace sme
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
+} // namespace pysme
