@@ -1,34 +1,34 @@
 #pragma once
 
-#include "sme_common.hpp"
 #include "sme_reaction.hpp"
 #include "sme_species.hpp"
-#include <pybind11/pybind11.h>
+#include <nanobind/nanobind.h>
 #include <string>
 #include <vector>
 
-namespace model {
+namespace sme::model {
 class Model;
 }
 
-namespace sme {
+namespace pysme {
 
-void pybindMembrane(pybind11::module &m);
+void bindMembrane(nanobind::module_ &m);
 
 class Membrane {
 private:
-  model::Model *s;
+  ::sme::model::Model *s;
   std::string id;
 
 public:
   Membrane() = default;
-  explicit Membrane(model::Model *sbmlDocWrapper, const std::string &sId);
+  explicit Membrane(::sme::model::Model *sbmlDocWrapper,
+                    const std::string &sId);
   [[nodiscard]] std::string getName() const;
   void setName(const std::string &name);
   std::vector<Reaction> reactions;
   [[nodiscard]] std::string getStr() const;
 };
 
-} // namespace sme
+} // namespace pysme
 
-PYBIND11_MAKE_OPAQUE(std::vector<sme::Membrane>)
+NB_MAKE_OPAQUE(std::vector<pysme::Membrane>)
