@@ -1,14 +1,15 @@
-// Python.h (included by pybind11.h) must come first
+// Python.h (#included by nanobind.h) must come first
 // https://docs.python.org/3.2/c-api/intro.html#include-files
-#include <pybind11/pybind11.h>
+#include <nanobind/nanobind.h>
 
 #include "sme/version.hpp"
 #include "sme_module.hpp"
 #include <QFile>
+#include <nanobind/stl/string.h>
 
-namespace sme {
+namespace pysme {
 
-void pybindModule(pybind11::module &m) {
+void bindModule(nanobind::module_ &m) {
   m.doc() = R"(
             Spatial Model Editor Python interface
 
@@ -17,7 +18,7 @@ void pybindModule(pybind11::module &m) {
 
             https://spatial-model-editor.readthedocs.io/
             )";
-  m.def("open_file", openFile, pybind11::arg("filename"),
+  m.def("open_file", openFile, nanobind::arg("filename"),
         R"(
         opens a sme or SBML file containing a spatial model
 
@@ -27,7 +28,7 @@ void pybindModule(pybind11::module &m) {
         Returns:
             Model: the spatial model
         )");
-  m.def("open_sbml_file", openSbmlFile, pybind11::arg("filename"),
+  m.def("open_sbml_file", openSbmlFile, nanobind::arg("filename"),
         R"(
         opens an SBML file containing a spatial model
 
@@ -38,7 +39,7 @@ void pybindModule(pybind11::module &m) {
             Model: the spatial model
         )");
   m.def("open_example_model", openExampleModel,
-        pybind11::arg("name") = "very-simple-model",
+        nanobind::arg("name") = "very-simple-model",
         R"(
         opens a built in example spatial model
 
@@ -79,7 +80,7 @@ void pybindModule(pybind11::module &m) {
                - Outside <-> Cell
                - Cell <-> Nucleus
         )");
-  m.attr("__version__") = common::SPATIAL_MODEL_EDITOR_VERSION;
+  m.attr("__version__") = ::sme::common::SPATIAL_MODEL_EDITOR_VERSION;
 }
 
 Model openFile(const std::string &filename) { return Model(filename); }
@@ -93,38 +94,13 @@ Model openExampleModel(const std::string &name) {
       f.open(QIODevice::ReadOnly | QIODevice::Text)) {
     xml = f.readAll().toStdString();
   } else {
-    throw SmeInvalidArgument("Failed to open example model '" + name +
-                             "'. Type help(sme.open_example_model) to see the "
-                             "available built-in models.");
+    throw std::invalid_argument(
+        "Failed to open example model '" + name +
+        "'. Type help(sme.open_example_model) to see the "
+        "available built-in models.");
   }
   m.importSbmlString(xml);
   return m;
 }
 
-} // namespace sme
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
-
-//
+} // namespace pysme
