@@ -1,32 +1,33 @@
 #pragma once
 
 #include "sme_reactionparameter.hpp"
-#include <pybind11/pybind11.h>
+#include <nanobind/nanobind.h>
 #include <string>
 #include <vector>
 
-namespace model {
+namespace sme::model {
 class Model;
 }
 
-namespace sme {
+namespace pysme {
 
-void pybindReaction(pybind11::module &m);
+void bindReaction(nanobind::module_ &m);
 
 class Reaction {
 private:
-  model::Model *s;
+  ::sme::model::Model *s;
   std::string id;
 
 public:
   Reaction() = default;
-  explicit Reaction(model::Model *sbmlDocWrapper, const std::string &sId);
+  explicit Reaction(::sme::model::Model *sbmlDocWrapper,
+                    const std::string &sId);
   [[nodiscard]] std::string getName() const;
   void setName(const std::string &name);
   std::vector<ReactionParameter> parameters;
   [[nodiscard]] std::string getStr() const;
 };
 
-} // namespace sme
+} // namespace pysme
 
-PYBIND11_MAKE_OPAQUE(std::vector<sme::Reaction>)
+NB_MAKE_OPAQUE(std::vector<pysme::Reaction>)
